NEW: A sample configuration.nix file is uploaded
The configuration file configuration_iso.nix is a sample of a basic configuration for generating the ISO image for a NixOS boot from USB. When downloaded it should be saved as configuration.nix. It contains the basic information including a regular user account, an account with sudo permissions, and a network availability. Additional/alternative packages can be added depending on requirements.