NEW: A sample of a change in the configuration
The configuration_rebuild.nix file is an example of if you would like to make some additional changes after installing the bootable NixOS. The example includes making the regular user account (in this case cobic) a temporary filesystem, and setting the time zone. It replaces the configuration_install.nix file before being updated.